Added API Versioning and Swagger Support
